move data helpers out of data directory
