+++ brainscore_vision/data/cadena2017/__init__.py
@@ -3,7 +3,7 @@ import logging
 from brainio.assemblies import NeuronRecordingAssembly
 from brainscore_vision import data_registry
 from brainscore_vision.utils.s3 import load_assembly_from_s3, load_stimulus_set_from_s3
-from brainscore_vision.data.data_helpers.helper import version_id_df, build_filename
+from brainscore_vision.data_helpers.helper import version_id_df, build_filename
 
 # extract version ids from version_ids csv
 assembly_version = version_id_df.at[build_filename('tolias.Cadena2017', '.nc'), 'version_id']
