+++ brainscore_vision/data/david2004/__init__.py
@@ -2,7 +2,7 @@ from brainio.assemblies import NeuronRecordingAssembly
 
 from brainscore_vision import data_registry, stimulus_set_registry
 from brainscore_vision.utils.s3 import load_assembly_from_s3, load_stimulus_set_from_s3
-from brainscore_vision.data.data_helpers.helper import version_id_df, build_filename
+from brainscore_vision.data_helpers.helper import version_id_df, build_filename
 
 BIBTEX = """@article{david2004evaluation,
   title={Evaluation of different measures of functional connectivity using a neural mass model},
